Updated Connection Interfaces in accordance with latest DR/spec changes
